requirements.txt
certifi==2020.6.20
click==8.0.4
Flask==2.0.3
greenlet==1.1.2
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
numpy==1.22.3
pandas==1.4.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.32
Werkzeug==2.0.3
